Improvement/remove warnings (#93)
* Update CHANGELOG.md

* Update CHANGELOG.md

* chore(release): 8.1.0 [skip ci]

# [8.1.0](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/compare/v8.0.2...v8.1.0) (2021-05-26)

### Bug Fixes

* **add SAVideo laod:** now accepts more arguments ([#85](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/issues/85)) ([1361b01](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/commit/1361b01171092d1b1ac6450fbe79f7d3f13bb436))

### Features

* **CI:** Fixed issues in gradle build scripts ([c076fed](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/commit/c076fed51d9c4890302206bc16af68771fa155fe))

* chore(release): 8.1.1 [skip ci]

## [8.1.1](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/compare/v8.1.0...v8.1.1) (2021-05-26)

### Bug Fixes

* **CI:** Releasing fix version ([01a41b4](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/commit/01a41b4a48642c6dc1ff0b6e509cf9472087b719))

* chore(release): 8.1.2 [skip ci]

## [8.1.2](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/compare/v8.1.1...v8.1.2) (2021-05-26)

### Bug Fixes

* **CI:** Adjusted CI config for tag ([ad24114](https://github.com/SuperAwesomeLTD/sa-mobile-sdk-android/commit/ad24114761c7e36d21c2520b36254a2816f1becf))

* remove more warnings

* remove onclick from xml

* use primatives

* remove hard coded string

+++ superawesome-base/src/main/java/tv/superawesome/lib/samodelspace/saad/SAAd.java
@@ -193,7 +193,7 @@ public class SAAd extends SABaseObject implements Parcelable {
 
         int val1 = SAJsonParser.getInt(jsonObject, "moat", (int) moat * 100);
         double val2 = SAJsonParser.getDouble(jsonObject, "moat", moat);
-        moat = Math.max(val2, (double) val1);
+        moat = Math.max(val2, val1);
         moat = moat > 1 ? 1 : moat;
 
         lineItemId = SAJsonParser.getInt(jsonObject, "line_item_id", lineItemId);
